fix: add rewind-blockchain command
